requirements.txt
# 腾讯云SDK依赖
tencentcloud-sdk-python-tiia
